ref: moved repo to profile, renamed project

+++ migrations/20220405151830_create_stage_table.up.sql
@@ -3,6 +3,6 @@ CREATE TABLE IF NOT EXISTS stage
     id         serial                NOT NULL PRIMARY KEY,
     name       varchar(32)           NOT NULL,
     is_final   boolean DEFAULT FALSE NOT NULL,
-    "order"    int                   NOT NULL,
+    "order"    smallint                   NOT NULL,
     package_id int                   NOT NULL REFERENCES package (id)
 );
